# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI CFO API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface DemoModeStatus {
  demoMode: boolean;
}

export interface SetDemoModeBody {
  demoMode: boolean;
}

export interface ConnectStripeBody {
  apiKey: string;
}

export interface StripeStatus {
  connected: boolean;
  accountName?: string;
}

export interface DashboardMetrics {
  /** Monthly Recurring Revenue in dollars */
  mrr: number;
  /** Month-over-month MRR growth percentage */
  mrrGrowthRate: number;
  /** Total revenue this month in dollars */
  totalRevenue: number;
  activeCustomers: number;
  activeSubscriptions: number;
  /** Monthly churn rate percentage */
  churnRate: number;
  avgRevenuePerUser: number;
  totalInvoices: number;
  overdueInvoices: number;
}

export interface RevenueDataPoint {
  /** e.g. "Jan 2025" */
  month: string;
  revenue: number;
  mrr: number;
  customers: number;
}

export interface RevenueChartResponse {
  data: RevenueDataPoint[];
}

export interface ForecastMonth {
  month: string;
  projectedRevenue: number;
  confidence: string;
}

export type InsightsResponseForecast = {
  summary: string;
  months: ForecastMonth[];
};

export type InsightsResponseRisks = {
  summary: string;
  items: string[];
};

export type InsightsResponseOpportunities = {
  summary: string;
  items: string[];
};

export type InsightsResponseRecommendedActions = {
  summary: string;
  items: string[];
};

export interface InsightsResponse {
  forecast: InsightsResponseForecast;
  risks: InsightsResponseRisks;
  opportunities: InsightsResponseOpportunities;
  recommendedActions: InsightsResponseRecommendedActions;
  generatedAt: string;
}

export type WeeklyActionsResponseActionsItem = {
  priority: number;
  action: string;
  rationale: string;
};

export interface WeeklyActionsResponse {
  actions: WeeklyActionsResponseActionsItem[];
  generatedAt: string;
}