# lib/api-zod/src/generated/types/dashboardMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI CFO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardMetrics {
  /** Monthly Recurring Revenue in dollars */
  mrr: number;
  /** Month-over-month MRR growth percentage */
  mrrGrowthRate: number;
  /** Total revenue this month in dollars */
  totalRevenue: number;
  activeCustomers: number;
  activeSubscriptions: number;
  /** Monthly churn rate percentage */
  churnRate: number;
  avgRevenuePerUser: number;
  totalInvoices: number;
  overdueInvoices: number;
}