# lib/api-zod/src/generated/types/insightsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI CFO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InsightsResponseForecast } from "./insightsResponseForecast";
import type { InsightsResponseOpportunities } from "./insightsResponseOpportunities";
import type { InsightsResponseRecommendedActions } from "./insightsResponseRecommendedActions";
import type { InsightsResponseRisks } from "./insightsResponseRisks";

export interface InsightsResponse {
  forecast: InsightsResponseForecast;
  risks: InsightsResponseRisks;
  opportunities: InsightsResponseOpportunities;
  recommendedActions: InsightsResponseRecommendedActions;
  generatedAt: Date;
}