# lib/api-zod/src/generated/types/stripeStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI CFO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StripeStatus {
  connected: boolean;
  accountName?: string;
}