# lib/api-zod/src/generated/types/weeklyActionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI CFO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyActionsResponseActionsItem } from "./weeklyActionsResponseActionsItem";

export interface WeeklyActionsResponse {
  actions: WeeklyActionsResponseActionsItem[];
  generatedAt: Date;
}